+++ vulkanalia/src/vk/mod.rs
@@ -7,6 +7,8 @@ mod builders;
 #[rustfmt::skip]
 mod commands;
 #[rustfmt::skip]
+mod enums;
+#[rustfmt::skip]
 mod extensions;
 #[rustfmt::skip]
 mod versions;
@@ -15,5 +17,6 @@ pub use vulkanalia_sys::*;
 
 pub use self::builders::*;
 pub use self::commands::*;
+pub use self::enums::*;
 pub use self::extensions::*;
 pub use self::versions::*;